Tweaks

+++ builtins.c
@@ -1552,7 +1552,7 @@ int dish_chmod(char **args)
             {
                 if (args[i + 1][j] >= '0' && args[i + 1][j] <= '7')
                 {
-                    mode += (args[i + 1][j] - '0') * pow(8, 2 - j);
+                    mode += (args[i + 1][j] - '0') * pow((double)8, (double)(2 - j));
                 } else
                 {
                     char *msg = malloc(sizeof(char) * MSG_LENGTH);
